moved redirect functionality to common and added unit tests
